Make middleware consistent
Middleware functions now all have the same signature - a function that
returns a function. This is a standard adopted across the Express
community and it should ease the transition if we switch out to
third-party modules for any of these.

I've also removed the requireUserAgent middleware, as this was an
unnecessary abstraction.