Add a user system (#69)
* Add a setup step (#62)

* Add a setup step

This configures the admin user and default permissions.

* Change the button text based on feedback

* Add in session support (#64)

* Add in session support

* Fix some lint errors introduced by the last PR

* Add authentication (#65)

* Add the ability to log in and out

* Split the demo and test seed data

This gives us more flexibility, and stops unrealistic test data from
making our demo data confusing. My intention is to add multiple sets of
test seed data for each test that we run, so that we know exactly where
the database is at for each test run.

* Overhaul the integration tests (#66)

This simplifies the integration tests and allows us to more easily load
different seed data for different tests. The seed data is also simpler
now that we can do this, because we don't need a single database to
cater for all possible test cases.

* Write integration tests for the setup step (#67)

* Write integration tests for the setup step

This ensures that the setup step is working properly.

* Fix some lint errors

* Add integration tests for login/logout (#68)

* Remove destructuring assignment 😬

* Use jsdom 9.x